Flask/restAPI/requirements.txt
flask
flask-smorest
flask-sqlalchemy
flask-migrate
flask-jwt-extended
marshmallow
passlib[bcrypt]
pymysql
